+++ flask_sqlalchemy_session2024/__init__.py
@@ -6,27 +6,15 @@ Flask-SQLAlchemy-Session
 Provides an SQLAlchemy scoped session that creates
 unique sessions per Flask request
 """
-# pylint: disable=invalid-name
 from werkzeug.local import LocalProxy
-from flask import _app_ctx_stack, current_app
+from flask import current_app, g
 from sqlalchemy.orm import scoped_session
 
 __all__ = ["current_session", "flask_scoped_session"]
 
 
 def _get_session():
-    # pylint: disable=missing-docstring, protected-access
-    context = _app_ctx_stack.top
-    if context is None:
-        raise RuntimeError(
-            "Cannot access current_session when outside of an application "
-            "context.")
-    app = current_app._get_current_object()
-    if not hasattr(app, "scoped_session"):
-        raise AttributeError(
-            "{0} has no 'scoped_session' attribute. You need to initialize it "
-            "with a flask_scoped_session.".format(app))
-    return app.scoped_session
+    return flask_scoped_session.get_scoped_session_from_flask()
 
 
 current_session = LocalProxy(_get_session)
@@ -41,35 +29,47 @@ class flask_scoped_session(scoped_session):
     """A :class:`~sqlalchemy.orm.scoping.scoped_session` whose scope is set to
     the Flask application context.
     """
+
+    flask_app_context_attribute = "_sqlalchemy_scoped_session"
+
     def __init__(self, session_factory, app=None):
         """
-        :param session_factory: A callable that returns a
-            :class:`~sqlalchemy.orm.session.Session`
-        :param app: a :class:`~flask.Flask` application
+        :param session_factory: A callable that returns a :class:`~sqlalchemy.orm.session.Session`
+        :param app: A :class:`~flask.Flask` application to be automatically registered with init_app
         """
-        super(flask_scoped_session, self).__init__(
-            session_factory,
-            scopefunc=_app_ctx_stack.__ident_func__)
-        # the _app_ctx_stack.__ident_func__ is the greenlet.get_current, or
-        # thread.get_ident if no greenlets are used.
-        # each Flask request is launched in a seperate greenlet/thread, so our
-        # session is unique per request
-        # _app_ctx_stack looks like internal API but is the only way to get to
-        # the active application context without adding logic to figure out
-        # whether threads, greenlets, or something else is used to create new
-        # application contexts. Keep in mind to refactor if Flask changes its
-        # public/private API towards this.
+        super().__init__(session_factory, self.__scopefunc)
+        self.__apps = set()
         if app is not None:
             self.init_app(app)
 
+    def __scopefunc(self):
+        # Check that the app has been registered. See "init_app()" for why.
+        assert (
+            current_app._get_current_object() in self.__apps
+        ), "you must register the app with flask_scoped_session.init_app(app) before using it"
+
+        return g._get_current_object()
+
     def init_app(self, app):
-        """Setup scoped session creation and teardown for the passed ``app``.
+        """
+        Install an appcontext teardown handler which closes the current scoped session. The
+        SQLAlchemy documentation warns that failing to call scoped_session.remove() can
+        result in resource leaks.
+
+        https://docs.sqlalchemy.org/en/20/orm/contextual.html#using-custom-created-scopes
 
         :param app: a :class:`~flask.Flask` application
         """
-        app.scoped_session = self
+        attr = self.flask_app_context_attribute
 
         @app.teardown_appcontext
         def remove_scoped_session(*args, **kwargs):
-            # pylint: disable=missing-docstring,unused-argument,unused-variable
-            app.scoped_session.remove()
+            self.remove()
+
+        # Permanently associate the app with this flask_scoped_session.
+        setattr(app, attr, self)
+        self.__apps.add(app)
+
+    @classmethod
+    def get_scoped_session_from_flask(cls):
+        return getattr(current_app, cls.flask_app_context_attribute)