+++ tests/test_flask_sqlalchemy_session.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
-import pytest
 try:
     from unittest import mock
 except ImportError:
     import mock
+
+import pytest
 from flask import Flask
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session, sessionmaker